pyproject.toml
[project]
name = "brasileirao-bi"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "python-dotenv==1.0.1",
  "requests==2.32.3",
  "pandas==2.2.2",
  "pyarrow==17.0.0",
  "google-cloud-bigquery==3.25.0",
  "google-auth==2.33.0",
]

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]